:fire: Remove 10px root font-size
This is a huge pet peeve of mine. Since px is a mostly-arbitrary unit nowadays, creating a layout system around 10px is an unnecessary practice. At least with the 62.5% hack, the font-size is accessible, but html { font-size: 10px } is not.

Instead, let our root continue to be 1rem (16px by default) and slightly lower our rem units to reflect this change